+++ lib/jazzy/source_declaration.rb
@@ -76,10 +76,14 @@ module Jazzy
       name.split(/[\(\)]/) if type.objc_category?
     end
 
-    def swift_extension_objc_name
-      return unless type.swift_extensible? || type.swift_extension?
+    def swift_objc_extension?
+      type.swift_extension? && usr && usr.start_with?('c:objc')
+    end
 
-      usr.split("(cs)").last
+    def swift_extension_objc_name
+      return unless type.swift_extension? && usr
+
+      usr.split('(cs)').last
     end
 
     # The language in the templates for display